Feat: followmodal 랜더링 문제 수정 by @ park74441
